requirements.txt
aiogram>=3.4.0
httpx[http2]>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
SQLAlchemy>=2.0.31
aiosqlite>=0.20.0
APScheduler>=3.10.4
python-dotenv>=1.0.1
